backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-google-genai==2.0.5

faiss-cpu==1.8.0.post1
numpy==1.26.4
tqdm==4.66.5
